+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     - `SingleCaseUnion`
     - `DiscriminatedUnion`
     - `Stream`
+- Parse/Check `Domain` of participants in tuc
+    - Component must have domain and it's participants must be in the same domain
 
 ## 0.2.0 - 2020-09-01
 - Parse left and right notes in tuc